+++ lib/money/currency.rb
@@ -12,6 +12,7 @@ class Money
   # @see http://iso4217.net/
   class Currency
     include Comparable
+    extend Enumerable
     extend Money::Currency::Loader
     extend Money::Currency::Heuristics
 
@@ -157,6 +158,12 @@ class Money
         existed ? true : false
       end
 
+
+      def each
+        all.each { |c| yield(c) }
+      end
+
+
       private
 
       def stringify_keys
@@ -240,7 +247,14 @@ class Money
     #   c2 <=> c1 #=> -1
     #   c1 <=> c1 #=> 0
     def <=>(other_currency)
-      self.priority <=> other_currency.priority
+      # <=> returns nil when one of the values is nil
+      comparison = self.priority <=> other_currency.priority || 0
+
+      if comparison == 0
+        self.id <=> other_currency.id
+      else
+        comparison
+      end
     end
 
     # Compares +self+ with +other_currency+ and returns +true+ if the are the